+++ tests/testthat/test-Kostka.R
@@ -6,3 +6,19 @@ test_that("Kostka numbers", {
     nssytx == sum(apply(compos, 2L, function(w) KostkaNumber(lambda, w)))
   )
 })
+
+test_that("Kostka numbers with given mu", {
+  mu <- c(3, 3, 2, 1)
+  kNumbers_mu <- KostkaNumbersWithGivenMu(mu, output = "vector")
+  lambdas <- partitions::parts(sum(mu))
+  kNumbers <- apply(lambdas, 2L, function(lambda) KostkaNumber(lambda, mu))
+  names(kNumbers) <- apply(lambdas, 2L, function(lambda) {
+    partitionAsString(removeTrailingZeros(lambda))
+  })
+  expect_true(
+    all(kNumbers[names(kNumbers_mu)] == kNumbers_mu)
+  )
+  expect_true(
+    all(kNumbers[setdiff(names(kNumbers), names(kNumbers_mu))] == 0L)
+  )
+})